Add apple pie recipe page and content
Link to another page for the apple pie
recipe keeps the homepage clean of clutter,
and organizes recipes for easy searching.

Image and description is provided so
the user knows what the expected outcome is
supposed to look and taste like. Ingredients are
listed so they can be bought in any order that is
most convenient for the user. Steps are numbered
for the user to follow instructions properly.